requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
pandas
python-dotenv
